GenAI/Production_Grade_System/requirements.txt
pyyaml==6.0.2
pydantic==2.9.2
requests==2.32.3
pytest==8.3.3

openai==1.40.6
anthropic==0.34.2

sentence-transformers==3.0.1
chromadb==0.5.5
